experiments/langchain-integration/requirements.txt
langchain-fireworks
python-dotenv 
numpy